Move pr event handlers in handler.js.
