+++ package.json
@@ -29,6 +29,7 @@
     "eslint-plugin-import": "^2.7.0",
     "mocha": "^4.0.1",
     "nyc": "^11.2.1",
+    "require-reload": "^0.2.2",
     "rewire": "^3.0.2",
     "sinon": "^4.0.1"
   }
